pyproject.toml
[project]
name = "chrono-marco2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hong", email = "qaz9756wsx@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "dxcam>=0.3.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pytelegrambotapi>=4.33.0",
    "scikit-image>=0.26.0",
]

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"
